Changed weights to decimal and added it to events

+++ src/ProductManagementAPI/Infrastructure/Events/ProductUpdated.cs
@@ -12,10 +12,10 @@ namespace ProductManagementAPI.Infrastructure.Events
 		public readonly string Id;
 		public readonly string Name;
 		public readonly decimal Price;
-		public readonly float WeightKg;
+		public readonly decimal WeightKg;
 		public readonly ProductType Type;
 
-		public ProductUpdated(Guid messageId, string id, string name, decimal price, float weightkg, ProductType type) :
+		public ProductUpdated(Guid messageId, string id, string name, decimal price, decimal weightkg, ProductType type) :
 			base(messageId, MessageTypes.ProductUpdated)
 		{
 			Id = id;
